Move settings stuff into a directory
